Chromium: Pass PageTitleTests
related to #72
